Massive version bump

+++ app/src/main/java/com/linkybook/comical/views/SiteEditor.java
@@ -18,20 +18,17 @@
 
 package com.linkybook.comical.views;
 
-import androidx.lifecycle.ViewModelProviders;
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.os.Bundle;
-import android.text.Editable;
-import android.text.TextWatcher;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 
-import com.linkybook.comical.SiteViewModel;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.lifecycle.ViewModelProviders;
+
 import com.linkybook.comical.R;
+import com.linkybook.comical.SiteViewModel;
 import com.linkybook.comical.data.SiteInfo;
 
 
@@ -57,42 +54,24 @@ public class SiteEditor extends AppCompatActivity {
 
         if(site == null) {
             site = new SiteInfo();
-
-            // Only adding on new sites so existing url doesn't get clobbered
-            name.addTextChangedListener(new TextWatcher() {
-                @Override
-                public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {}
-
-                @Override
-                public void onTextChanged(CharSequence charSequence, int i, int i1, int i2) {}
-
-                @Override
-                public void afterTextChanged(Editable editable) {
-                    String name = editable.toString().toLowerCase().replaceAll("[^a-z]", "");
-                    SiteEditor.this.url.setText("http://www." + name + ".com");
-                }
-            });
         } else {
             name.setText(site.name);
             url.setText(site.url);
             submitButton.setText(R.string.prompt_update);
         }
 
-        submitButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                site.name = name.getText().toString();
-                site.url = url.getText().toString();
-                svm.addOrUpdateSite(site);
-                finish();
-            }
+        submitButton.setOnClickListener(view -> {
+            site.name = name.getText().toString();
+            site.url = url.getText().toString();
+            svm.addOrUpdateSite(site);
+            finish();
         });
     }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         getMenuInflater().inflate(R.menu.edit_menu, menu);
-        if(site.favorite == true) {
+        if(site.favorite) {
             menu.findItem(R.id.action_favorite).setIcon(R.drawable.ic_favorite_black_24dp);
         }
         if(site.name != null) {
@@ -105,21 +84,18 @@ public class SiteEditor extends AppCompatActivity {
     public boolean onOptionsItemSelected(MenuItem item) {
         SiteInfo site = SiteEditor.this.site;
 
-        switch (item.getItemId()) {
-            case R.id.action_favorite:
-                site.favorite = !site.favorite;
-                if(site.favorite == true) {
-                    item.setIcon(R.drawable.ic_favorite_black_24dp);
-                } else {
-                    item.setIcon(R.drawable.ic_favorite_border_black_24dp);
-                }
-                break;
-            case R.id.action_delete:
+        if(item.getItemId() == R.id.action_favorite) {
+            site.favorite = !site.favorite;
+            if(site.favorite) {
+                item.setIcon(R.drawable.ic_favorite_black_24dp);
+            } else {
+                item.setIcon(R.drawable.ic_favorite_border_black_24dp);
+            }
+        } else if(item.getItemId() == R.id.action_delete) {
                 svm.deleteSite(site);
                 finish();
-                break;
-            default:
-                return super.onOptionsItemSelected(item);
+        } else {
+            return super.onOptionsItemSelected(item);
         }
         return true;
     }
